// src/generated/pistonpanel/instance.ts
export interface InstanceListResponse {
  /**
   * @generated from protobuf field: repeated pistonpanel.v1.InstanceListResponse.Instance instances = 1;
   */
  instances: InstanceListResponse_Instance[];
}
/**
 * @generated from protobuf message pistonpanel.v1.InstanceListResponse.Instance
 */
export interface InstanceListResponse_Instance {
  /**
   * @generated from protobuf field: string id = 1;
   */
  id: string;
  /**
   * @generated from protobuf field: string friendly_name = 2;
   */
  friendlyName: string;
  /**
   * @generated from protobuf field: string icon = 5;
   */
  icon: string;
  /**
   * @generated from protobuf field: pistonpanel.v1.InstanceState state = 3;
   */
  state: InstanceState;
  /**
   * @generated from protobuf field: repeated pistonpanel.v1.InstancePermissionState instance_permissions = 4;
   */
  instancePermissions: InstancePermissionState[];
}
/**
 * @generated from protobuf message pistonpanel.v1.InstancePermissionState
 */
export interface InstancePermissionState {
  /**
   * @generated from protobuf field: pistonpanel.v1.InstancePermission instance_permission = 1;
   */
  instancePermission: InstancePermission;
  /**
   * @generated from protobuf field: bool granted = 2;
   */
  granted: boolean;
}
/**
 * @generated from protobuf message pistonpanel.v1.InstanceInfoRequest
 */
export interface InstanceInfoRequest {
  /**
   * @generated from protobuf field: string id = 1;
   */
  id: string;
}
/**
 * @generated from protobuf message pistonpanel.v1.InstanceInfoResponse
 */
export interface InstanceInfoResponse {
  /**
   * @generated from protobuf field: string friendly_name = 1;
   */
  friendlyName: string;
  /**
   * @generated from protobuf field: string icon = 5;
   */
  icon: string;
  /**
   * @generated from protobuf field: pistonpanel.v1.InstanceState state = 3;
   */
  state: InstanceState;
  /**
   * @generated from protobuf field: repeated pistonpanel.v1.InstancePermissionState instance_permissions = 4;
   */
  instancePermissions: InstancePermissionState[];
}
// ...
export enum InstanceState {
  /**
   * @generated from protobuf enum value: STARTING = 0;
   */
  STARTING = 0,
  /**
   * @generated from protobuf enum value: RUNNING = 1;
   */
  RUNNING = 1,
  /**
   * @generated from protobuf enum value: PAUSED = 2;
   */
  PAUSED = 2,
  /**
   * @generated from protobuf enum value: STOPPING = 3;
   */
  STOPPING = 3,
  /**
   * @generated from protobuf enum value: STOPPED = 4;
   */
  STOPPED = 4,
}
// ...
class InstanceListResponse$Type extends MessageType<InstanceListResponse> {
  constructor() {
    super('pistonpanel.v1.InstanceListResponse', [
      {
        no: 1,
        name: 'instances',
        kind: 'message',
        repeat: 1 /*RepeatType.PACKED*/,
        T: () => InstanceListResponse_Instance,
      },
    ]);
  }
}
/**
 * @generated MessageType for protobuf message pistonpanel.v1.InstanceListResponse
 */
export const InstanceListResponse = new InstanceListResponse$Type();
// @generated message type with reflection information, may provide speed optimized methods
class InstanceListResponse_Instance$Type extends MessageType<InstanceListResponse_Instance> {
  constructor() {
    super('pistonpanel.v1.InstanceListResponse.Instance', [
      { no: 1, name: 'id', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
      {
        no: 2,
        name: 'friendly_name',
        kind: 'scalar',
        T: 9 /*ScalarType.STRING*/,
      },
      { no: 5, name: 'icon', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
      {
        no: 3,
        name: 'state',
        kind: 'enum',
        T: () => ['pistonpanel.v1.InstanceState', InstanceState],
      },
      {
        no: 4,
        name: 'instance_permissions',
        kind: 'message',
        repeat: 1 /*RepeatType.PACKED*/,
        T: () => InstancePermissionState,
      },
    ]);
  }
}
/**
 * @generated MessageType for protobuf message pistonpanel.v1.InstanceListResponse.Instance
 */
export const InstanceListResponse_Instance =
  new InstanceListResponse_Instance$Type();
// @generated message type with reflection information, may provide speed optimized methods
class InstancePermissionState$Type extends MessageType<InstancePermissionState> {
  constructor() {
    super('pistonpanel.v1.InstancePermissionState', [
      {
        no: 1,
        name: 'instance_permission',
        kind: 'enum',
        T: () => ['pistonpanel.v1.InstancePermission', InstancePermission],
      },
      { no: 2, name: 'granted', kind: 'scalar', T: 8 /*ScalarType.BOOL*/ },
    ]);
  }
}
/**
 * @generated MessageType for protobuf message pistonpanel.v1.InstancePermissionState
 */
export const InstancePermissionState = new InstancePermissionState$Type();
// ...
class InstanceInfoResponse$Type extends MessageType<InstanceInfoResponse> {
  constructor() {
    super('pistonpanel.v1.InstanceInfoResponse', [
      {
        no: 1,
        name: 'friendly_name',
        kind: 'scalar',
        T: 9 /*ScalarType.STRING*/,
      },
      { no: 5, name: 'icon', kind: 'scalar', T: 9 /*ScalarType.STRING*/ },
      {
        no: 3,
        name: 'state',
        kind: 'enum',
        T: () => ['pistonpanel.v1.InstanceState', InstanceState],
      },
      {
        no: 4,
        name: 'instance_permissions',
        kind: 'message',
        repeat: 1 /*RepeatType.PACKED*/,
        T: () => InstancePermissionState,
      },
    ]);
  }
}
/**
 * @generated MessageType for protobuf message pistonpanel.v1.InstanceInfoResponse
 */
export const InstanceInfoResponse = new InstanceInfoResponse$Type();